fix(plasma-core): rework slider types
