+++ packages/plasma-core/src/components/Slider/Slider.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 
-import { Slider as SliderSingleValue, SliderProps as SingleSliderProps } from './Single';
-import { Slider as SliderDoubleValues, SliderProps as DoubleSliderProps } from './Double';
-
-export type SliderProps = SingleSliderProps | DoubleSliderProps;
+import { Slider as SliderSingleValue } from './Single';
+import { Slider as SliderDoubleValues } from './Double';
+import type { SliderProps, SingleSliderProps, SliderInternalProps } from './types';
 
 const isSingleValueProps = (props: SliderProps): props is SingleSliderProps => typeof props.value === 'number';
 
@@ -12,7 +11,7 @@ const isSingleValueProps = (props: SliderProps): props is SingleSliderProps => t
  * Можно указать два значения.
  * Только для приложения Салют.
  */
-export const SliderCore: React.FC<SliderProps> = (props) => {
+export const SliderCore: React.FC<SliderProps & SliderInternalProps> = (props) => {
     if (isSingleValueProps(props)) {
         return <SliderSingleValue {...props} />;
     }
