+++ packages/plasma-core/src/components/Slider/index.ts
@@ -3,9 +3,6 @@ import { SliderCore } from './Slider';
 export { SliderCore } from './Slider';
 export { ThumbBase } from './ThumbBase';
 
-export type { SliderProps } from './Slider';
-export type { SliderProps as SliderSingleProps } from './Single';
-export type { SliderProps as SliderDoubleProps } from './Double';
-export type { SliderBaseProps, SliderSettings } from './types';
+export type { SliderSettings, SliderProps, SingleSliderProps, DoubleSliderProps } from './types';
 
 export const Slider = SliderCore;
