+++ packages/plasma-core/src/components/Slider/types.ts
@@ -37,7 +37,7 @@ export type SliderSettings = Partial<{
     fillColor: string;
 }>;
 
-export interface SliderBaseProps {
+interface SliderBaseProps {
     /**
      * Минимальное значение
      */
@@ -50,9 +50,70 @@ export interface SliderBaseProps {
      * Компонент неактивен
      */
     disabled?: boolean;
+}
 
+export interface SliderInternalProps extends ThumbProp {
     /**
      * Настройки внешнего вида slider
      */
     settings?: SliderSettings;
 }
+
+export interface SliderViewProps extends SliderBaseProps, Omit<SliderInternalProps, 'thumb'> {
+    railFillWidth: number;
+    railFillXPosition?: number;
+    setStepSize(stepSize: number): void;
+    onChange?(value: number, data: DraggableData): void;
+}
+
+export interface SingleSliderProps extends SliderBaseProps {
+    /**
+     * Текущее значение
+     */
+    value: number;
+    /**
+     * Ярлык, определяющий назначение ползунка, например «Минимальная цена» [a11y].
+     */
+    ariaLabel?: string;
+    /**
+     * Размера увеличенного шага (для клавиш PageUp, PageDown).
+     * Указывает процентное отношение от максимально возможного значения.
+     * Указав значение 20 при максимуме в 100, получим 20%.
+     */
+    multipleStepSize?: number;
+    /**
+     * Вызывается при перемещении ползунка
+     */
+    onChangeCommitted(value: number): void;
+    /**
+     * Вызывается при отпускании ползунка
+     */
+    onChange?(value: number): void;
+}
+
+export interface DoubleSliderProps extends SliderBaseProps {
+    /**
+     * Текущее значение
+     */
+    value: number[];
+    /**
+     * Ярлык, определяющий назначение ползунка, например «Минимальная цена» [a11y].
+     */
+    ariaLabel?: string[];
+    /**
+     * Размера увеличенного шага (для клавиш PageUp, PageDown).
+     * Указывает процентное отношение от максимально возможного значения.
+     * Указав значение 20 при максимуме в 100, получим 20%.
+     */
+    multipleStepSize?: number;
+    /**
+     * Вызывается при отпускании ползунка
+     */
+    onChangeCommitted(value: number[]): void;
+    /**
+     * Вызывается при перемещении ползунка
+     */
+    onChange?(value: number[]): void;
+}
+
+export type SliderProps = SingleSliderProps | DoubleSliderProps;
